Fix Mario collision, Added Enemy collision
Fixed the script that were recognising Mario not falling because he was falling on his block. I even added the logic for the collision detection of the enemys so they now bounce whenever they should be hitting something. I havent added that they can fall but it can wait a little longer :D