Added end collision+SuperMario+Remove the powerUps
I added the collision with the end flag, and fixed it's animation, now it stops at the right block. Added the transformation of mario into SuperMario, and added how to remove the PowerUps.
We still have a big problem, that is SuperMario under hitboxes arent the rightones, this makes mario go under the blocks and not have the right collisions